feat: Refactor extension methods
